requirements.txt
base58==2.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colander==1.7.0
ecdsa==0.15
hemerton==0.1
idna==2.9
iso8601==0.1.12
libeospy==2.0.1
prompt-toolkit==1.0.14
Pygments==2.5.2
PyInquirer==1.0.3
pytz==2019.3
PyYAML==5.3
regex==2020.2.20
requests==2.23.0
six==1.14.0
termcolor==1.1.0
translationstring==1.3
typing-extensions==3.7.4.1
urllib3==1.25.8
wcwidth==0.1.8
